bot/requirements.txt
python-dateutil==2.8.2
Telethon==1.28.5
aiogram==3.6.0
miniopy-async==1.19
motor==3.4.0
python-dotenv==1.0.1
aiohttp==3.9.5
selenium==4.21.0
tqdm==4.66.4
bs4==0.0.2